Remove hardcoded TLS version for net46
Added a target for .NET Framework 4.7 which does not require hardcoding
the TLS version. This should prevent most issues as targetting .NET 4.6
and running on 4.7 causes issues with TLS versions not being selected by
the operating system.

If people are running on .NET 4.6 they will have to enable TLS 1.2
anyway, so we shouldn't be doing anything special there.

This change brings the library in line with the TLS best practices as
outlined by Microsoft.